refactor(chacha12-fast-check-test): adopt test-package shape
Move the fast-check / pure-rand integration helper to the
test-package convention used elsewhere in the workspace:

- Move `index.js` to `test/_random-type.js` (test packages export
  nothing; helper modules live in `test/` with a leading underscore).
- Drop `main` and `module` from `package.json`; reduce `exports` to
  just `./package.json`.
- Delete `README.md` (omissible on a test package).
- Delete the `packages/chacha12/test/index.test.js` placeholder stub.
